Add possibility to initialize only a given collection

+++ sources/metadata/metadata-core/src/main/java/fr/gouv/vitam/metadata/core/database/collections/MetadataCollections.java
@@ -27,14 +27,18 @@
 package fr.gouv.vitam.metadata.core.database.collections;
 
 import com.google.common.annotations.VisibleForTesting;
+import com.google.common.collect.Lists;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
+import fr.gouv.vitam.common.ParametersChecker;
 import fr.gouv.vitam.common.database.collections.VitamCollection;
 import fr.gouv.vitam.common.database.collections.VitamCollectionHelper;
 import fr.gouv.vitam.common.guid.GUIDFactory;
 import org.bson.Document;
 
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
 import java.util.List;
 
 /**
@@ -56,7 +60,15 @@ public enum MetadataCollections {
     @VisibleForTesting
     public static void beforeTestClass(final MongoDatabase db, String prefix,
         final ElasticsearchAccessMetadata esClient, Integer... tenants) {
-        for (MetadataCollections collection : MetadataCollections.values()) {
+        beforeTestClass(db, prefix, esClient, Lists.newArrayList(MetadataCollections.values()), tenants);
+    }
+
+    @VisibleForTesting
+    public static void beforeTestClass(final MongoDatabase db, String prefix,
+        final ElasticsearchAccessMetadata esClient, Collection<MetadataCollections> metadataCollections,
+        Integer... tenants) {
+        ParametersChecker.checkParameter("metadataCollections is required", metadataCollections);
+        for (MetadataCollections collection : metadataCollections) {
             collection.vitamCollection
                 .setName(prefix + collection.vitamCollection.getClasz().getSimpleName());
             collection.initialize(db, false);
@@ -69,11 +81,31 @@ public enum MetadataCollections {
 
     @VisibleForTesting
     public static void afterTestClass(final ElasticsearchAccessMetadata esClient, Integer... tenants) {
-        for (MetadataCollections collection : MetadataCollections.values()) {
+        afterTestClass(esClient, Lists.newArrayList(MetadataCollections.values()), false, tenants);
+    }
+
+    @VisibleForTesting
+    public static void afterTestClass(final ElasticsearchAccessMetadata esClient, boolean deleteEsIndex,
+        Integer... tenants) {
+        afterTestClass(esClient, Lists.newArrayList(MetadataCollections.values()), deleteEsIndex, tenants);
+    }
+
+    @VisibleForTesting
+    public static void afterTestClass(final ElasticsearchAccessMetadata esClient,
+        Collection<MetadataCollections> metadataCollections, boolean deleteEsIndex, Integer... tenants) {
+        if (null == metadataCollections) {
+            return;
+        }
+
+        for (MetadataCollections collection : metadataCollections) {
             collection.vitamCollection.getCollection().deleteMany(new Document());
 
             for (Integer tenant : tenants) {
-                esClient.purgeIndex(collection.getName(), tenant);
+                if (deleteEsIndex) {
+                    esClient.deleteIndex(collection, tenant);
+                } else {
+                    esClient.purgeIndex(collection.getName(), tenant);
+                }
             }
         }
     }
